requirements.txt
# Project Aether - Requirements
# AIOps Multi-Agent System

# Core Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# AI/LLM
ollama>=0.6.0

# Kubernetes & Infrastructure
kubernetes>=29.0.0
pyyaml>=6.0.1

# Knowledge Graph
neo4j>=5.15.0
networkx>=3.2.1

# Observability
prometheus-client>=0.19.0
requests>=2.31.0
httpx>=0.26.0

# Data Processing
pandas>=2.1.4
numpy>=1.26.3

# Testing & Chaos Engineering
chaostoolkit>=1.19.0
pytest>=7.4.4
pytest-asyncio>=0.23.3

# Utilities
click>=8.1.7
rich>=13.7.0
jinja2>=3.1.3